Split up analyzer defs from static analyzers #404
